Switched from BigDecimal to Zero

+++ src/main/kotlin/com/greenkode/trader/event/OrderEvent.kt
@@ -4,14 +4,13 @@ import com.greenkode.trader.domain.EventTypeEnum
 import com.greenkode.trader.domain.OrderDirection
 import com.greenkode.trader.domain.OrderType
 import com.greenkode.trader.domain.Symbol
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 
 class OrderEvent(
     val symbol: Symbol,
     val orderType: OrderType,
-    val quantity: BigDecimal,
+    val quantity: Double,
     val direction: OrderDirection,
     val timestamp: LocalDateTime
 ) : Event {
